+++ control-plane/aegisops_control_plane/action_review_projection.py
@@ -1,105 +1,9 @@
+"""Compatibility shim for the action review projection module."""
+
 from __future__ import annotations
 
-from .action_review_chain import (
-    _action_review_approval_decision,
-    _action_review_approval_state,
-    _action_review_execution,
-    _action_review_state,
-    _latest_action_review_reconciliation,
-    _next_expected_action_for_review_state,
-    _replacement_action_request,
-    action_review_chains_for_scope,
-    action_review_priority,
-    build_action_review_chain_snapshot,
-)
-from .action_review_coordination import (
-    action_review_coordination_ticket_mismatch,
-    action_review_coordination_ticket_outcome,
-    action_review_coordination_ticket_terminal_issue,
-    action_review_downstream_binding,
-)
-from .action_review_index import (
-    _ActionReviewRecordIndex,
-    _action_request_is_review_bound,
-    build_action_review_record_index,
-)
-from .action_review_path_health import (
-    _action_review_delegation_path_health,
-    _action_review_ingest_path_health,
-    _action_review_overall_path_state,
-    _action_review_overdue_path_health,
-    _action_review_path_health_entries,
-    _action_review_path_health_entry,
-    _action_review_path_health_summary,
-    _action_review_persistence_path_health,
-    _action_review_provider_path_health,
-    _action_review_reconciliation_without_execution_path_health,
-    _action_review_terminal_non_delegated_path_health,
-    _action_review_unresolved_without_execution_path_health,
-    _action_review_visibility_deadline,
-    action_review_path_health,
-)
-from .action_review_timeline import (
-    _action_review_stage_snapshot,
-    action_review_mismatch_inspection,
-    action_review_reconciliation_detail,
-    action_review_timeline,
-)
-from .action_review_visibility import (
-    _action_review_after_hours_handoff_visibility,
-    _action_review_context_matches_lineage,
-    _action_review_escalation_visibility,
-    _action_review_manual_fallback_visibility,
-    _action_review_visibility_context,
-    _action_review_visibility_entry,
-    _action_review_visibility_update,
-    _case_has_single_review_bound_action_request,
-    action_review_runtime_visibility,
-)
+import sys
 
-__all__ = [
-    "_ActionReviewRecordIndex",
-    "_action_request_is_review_bound",
-    "_action_review_after_hours_handoff_visibility",
-    "_action_review_approval_decision",
-    "_action_review_approval_state",
-    "_action_review_context_matches_lineage",
-    "_action_review_delegation_path_health",
-    "_action_review_escalation_visibility",
-    "_action_review_execution",
-    "_action_review_ingest_path_health",
-    "_action_review_manual_fallback_visibility",
-    "_action_review_overall_path_state",
-    "_action_review_overdue_path_health",
-    "_action_review_path_health_entries",
-    "_action_review_path_health_entry",
-    "_action_review_path_health_summary",
-    "_action_review_persistence_path_health",
-    "_action_review_provider_path_health",
-    "_action_review_reconciliation_without_execution_path_health",
-    "_action_review_stage_snapshot",
-    "_action_review_state",
-    "_action_review_terminal_non_delegated_path_health",
-    "_action_review_unresolved_without_execution_path_health",
-    "_action_review_visibility_context",
-    "_action_review_visibility_deadline",
-    "_action_review_visibility_entry",
-    "_action_review_visibility_update",
-    "_case_has_single_review_bound_action_request",
-    "_latest_action_review_reconciliation",
-    "_next_expected_action_for_review_state",
-    "_replacement_action_request",
-    "action_review_chains_for_scope",
-    "action_review_coordination_ticket_mismatch",
-    "action_review_coordination_ticket_outcome",
-    "action_review_coordination_ticket_terminal_issue",
-    "action_review_downstream_binding",
-    "action_review_mismatch_inspection",
-    "action_review_path_health",
-    "action_review_priority",
-    "action_review_reconciliation_detail",
-    "action_review_runtime_visibility",
-    "action_review_timeline",
-    "build_action_review_chain_snapshot",
-    "build_action_review_record_index",
-]
+from .actions.review import action_review_projection as _impl
+
+sys.modules[__name__] = _impl
