+++ control-plane/aegisops_control_plane/actions/review/__init__.py
@@ -1,1 +1,1 @@
-"""Package scaffold only for future reviewed-action modules."""
+"""Package scaffold only marker retained for reviewed-action surface modules."""
