+++ scripts/verify-phase-52-5-2-import-compatibility.sh
@@ -23,6 +23,24 @@ required_imports=(
   "aegisops_control_plane.reporting.audit_export:export_audit_retention_baseline"
   "aegisops_control_plane.pilot_reporting_export:export_pilot_executive_summary"
   "aegisops_control_plane.reporting.pilot_reporting_export:export_pilot_executive_summary"
+  "aegisops_control_plane.action_review_chain:action_review_chains_for_scope"
+  "aegisops_control_plane.actions.review.action_review_chain:action_review_chains_for_scope"
+  "aegisops_control_plane.action_review_coordination:action_review_coordination_ticket_outcome"
+  "aegisops_control_plane.actions.review.action_review_coordination:action_review_coordination_ticket_outcome"
+  "aegisops_control_plane.action_review_index:build_action_review_record_index"
+  "aegisops_control_plane.actions.review.action_review_index:build_action_review_record_index"
+  "aegisops_control_plane.action_review_inspection:ActionReviewInspectionBoundary"
+  "aegisops_control_plane.actions.review.action_review_inspection:ActionReviewInspectionBoundary"
+  "aegisops_control_plane.action_review_path_health:action_review_path_health"
+  "aegisops_control_plane.actions.review.action_review_path_health:action_review_path_health"
+  "aegisops_control_plane.action_review_projection:build_action_review_record_index"
+  "aegisops_control_plane.actions.review.action_review_projection:build_action_review_record_index"
+  "aegisops_control_plane.action_review_timeline:action_review_timeline"
+  "aegisops_control_plane.actions.review.action_review_timeline:action_review_timeline"
+  "aegisops_control_plane.action_review_visibility:action_review_runtime_visibility"
+  "aegisops_control_plane.actions.review.action_review_visibility:action_review_runtime_visibility"
+  "aegisops_control_plane.action_review_write_surface:ActionReviewWriteSurface"
+  "aegisops_control_plane.actions.review.action_review_write_surface:ActionReviewWriteSurface"
 )
 
 if [[ ! -f "${doc_path}" ]]; then
